+++ packages/gamemode/src/game/sagas.ts
@@ -1,6 +1,6 @@
+import { Logger } from "winston";
 import { GameOptions } from "@creature-chess/models";
 import { Player } from "../player";
-import { IOpponentProvider } from "./opponentProvider";
 
 export type GameSagaDependencies = {
     getMatchups: () => { homeId: string, awayId: string, awayIsClone: boolean }[];
@@ -9,5 +9,6 @@ export type GameSagaDependencies = {
         getAll: () => Player[];
         getLiving: () => Player[];
         getById: (id: string) => Player | null;
-    }
+    };
+    logger: Logger;
 };