+++ packages/auth-server/CHANGELOG.md
@@ -3,6 +3,28 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.3.46](https://github.com/Jameskmonger/creature-chess/compare/v0.3.44...v0.3.46) (2021-04-21)
+
+
+### Features
+
+* store nickname against auth0 user ([4c35183](https://github.com/Jameskmonger/creature-chess/commit/4c351839885fa33c4a99b52cb1fc7a0a1ac2e4c6))
+
+
+
+
+
+## [0.3.45](https://github.com/Jameskmonger/creature-chess/compare/v0.3.44...v0.3.45) (2021-04-21)
+
+
+### Features
+
+* store nickname against auth0 user ([f1b7d45](https://github.com/Jameskmonger/creature-chess/commit/f1b7d45ee06370678e165c8e06835b01adc3ef2b))
+
+
+
+
+
 ## [0.3.44](https://github.com/Jameskmonger/creature-chess/compare/v0.3.42...v0.3.44) (2021-04-16)
 
 **Note:** Version bump only for package @creature-chess/auth-server